refactor(RBBalancerTest): Move the RBT model to a separate class and separate insert/delete test cases
